quotes/requirements.txt
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.0
